refactor(v5): move results publishing out of the god class (#878, 4/4)
Final extraction. BaseTerraformCommandHandler goes from 1750 to 1369 lines --
2302 when this series started, a 40% reduction across the four PRs.

src/results-publisher.ts owns everything that turns a finished command's output
into something a pipeline can see: the structured digest attachments (#881), the
TF_OUT_* variables behind the printable-ASCII guard, the apply -json console
echo with its ##vso neutralization (#678), the 0600 command-output files (#484),
and the sensitive-output warn/fail checks (#488, #650, #802). It takes the
TempFileManager because two of those checks register the cleartext file they
just inspected for scrub+delete before failing the task -- a strict failure must
not leave the values on disk.

publishPlanSummaryAttachment and publishStateSummaryAttachment did NOT come
along, despite the names. Both shell out to `terraform show` before they publish
anything, so moving them would have meant injecting the tool handler, the
command executor and createBaseCommand into a results class. Needing three
orchestration capabilities is a sign the seam is in the wrong place, so they
stay with the per-command code. getTaskVersion and
relativeWorkingDirectoryForDigest DID move -- they exist only to fill
buildDigestMeta's provenance fields, and leaving them behind would have made the
new module import the handler it was just extracted from.

The risk in this one was in tests the compiler cannot check. Three class tests
assert against the handler's SOURCE TEXT, read by literal path, so moved code
breaks them by string rather than by type. Two were path updates. The third was
not:

    const bareWrites = base.match(/fs\.writeFileSync\(/g) || [];
    assert.strictEqual(bareWrites.length, 0, ...);

The handler no longer imports fs at all, so after the move that assertion would
have passed trivially and permanently while the write sinks it exists to guard
had relocated to a file it never reads -- a green tick certifying nothing. It
now reads both files. The SITE_ROWS audit table needed the same treatment: its
loop only asserts fs.existsSync(file), so two stale rows would have kept passing
while pointing at a file that no longer contains the function.

Two self-inflicted breaks, both caught by the real build rather than by editor
diagnostics, both worth recording. An import-block rewrite left a duplicate
`import os` because the original sat just outside the replaced span -- two
TS2300s, suite aborted, zero tests run, and get_errors had called the file
clean. And the strengthened assertion above first landed as a .concat() over
`never[] | RegExpMatchArray`, which does not typecheck; it is a spread now.

Ten imports went dead here. `fs` was the interesting one: a plain grep finds it
"used" at line 823, but that is inside a comment -- the name survives only in
prose, which is exactly the shape that failed lint on #920. The
comment-stripping check caught it; a grep would not have.

Gate: 9/9 repo gates exit 0, lint 0, 786 passing / 0 failing, per-file floor
met. results-publisher.js 87.94 stmts / 85.18 branch / 86.2 func / 91.04 lines
against a floor of 60; the handler holds at 91.59 lines. Test count unchanged
across all four PRs.

What is left in the handler is the per-command implementations themselves plus
provider-detection output parsing. Both are worth doing, but neither is a lift:
the commands are the class's actual reason to exist, and splitting them further
means deciding whether the provider subclass hierarchy survives. That is a
design question, not a mechanical move, so it is not being smuggled into this
series.
