+++ CLAUDE.md
@@ -156,6 +156,7 @@ azure-pipelines-terraform/
 | `temp-file-manager.ts`                 | Owns the command handler's temp-file lifecycle: the two tracking tiers (normal-completion vs cancellation-only, #650), scrub-then-unlink of each tracked path (#595), and the secure var file's scrub-before-delete (#662)                                                                                                                   |
 | `argument-builder.ts`                  | Turns task inputs into Terraform argv: the `commandOptions` tokenizer that mirrors task-lib's `_argStringToArray` (#875) and the readers built on it (`-out=`, `-json`, positional plan file), the multi-line input parsers (`-var-file`, `-target`, `-var`), and `ArgumentBuilder` for the input-reading flags (`-replace`, `-parallelism`, secure var file) |
 | `command-executor.ts`                  | Runs a `ToolRunner` and turns the result into output or an error: the wall-clock deadline (#822), the bounded stdout/stderr capture that forces `silent: true` (#492, #613, #632), fmt/test's separate message capture (#826), and the loc-keyed failure formatter (#821)                                                                                     |
+| `results-publisher.ts`                 | Turns a finished command's output into what a pipeline consumes: structured digest attachments (#881), `TF_OUT_*` variables behind the printable-ASCII guard, the `apply -json` console echo with `##vso` neutralization (#678), 0600 command-output files (#484), and the sensitive-output warn/fail checks (#488, #650, #802)                              |
 | `secure-file-loader.ts`                | Downloads secure var files from ADO Secure Files library                                                                                                                                                                                                                                                                                     |
 | `id-token-generator.ts`                | Generates OIDC ID tokens for Workload Identity Federation fallback                                                                                                                                                                                                                                                                           |
 | `secure-temp.ts`                       | Secure temp-file primitives: owner-only 0600 + O_EXCL on Unix, a restrictive icacls DACL on Windows (both fail closed), plus symlink-guarded `scrubFile()` zero-overwrite-before-unlink (#595) — canonical source; byte-identical copy also in TerraformDriftReportV1 and TerraformPolicyCheckV1, gated by `scripts/check-shared-modules.js` |
